+++ src/Plainion.Windows.Tests/Controls/Text/AutoCorrection/ClickableHyperlinkTests.cs
@@ -1,0 +1,28 @@
+﻿using System.Linq;
+using System.Windows.Documents;
+using NUnit.Framework;
+using Plainion.Windows.Controls.Text;
+using Plainion.Windows.Controls.Text.AutoCorrection;
+
+namespace Plainion.Windows.Tests.Controls.Text.AutoCorrection
+{
+    [TestFixture]
+    class ClickableHyperlinkTests
+    {
+        [Test]
+        public void TryMakeHyperlinks_AfterLink_HyperlinkInserted([Values("http://github.com/", "https://github.com/", "ftp://github.com/")]string url)
+        {
+            var document = new FlowDocument(new Paragraph(new Run("Some dummy " + url)));
+
+            new ClickableHyperlink().TryApply(new TextRange(document.ContentEnd, document.ContentEnd));
+
+            var visitor = new FlowDocumentVisitor(e => e is Hyperlink);
+            visitor.Accept(document);
+            Assert.That(visitor.Results.Count, Is.EqualTo(1));
+
+            var hyperlink = visitor.Results.OfType<Hyperlink>().Single();
+            Assert.That(hyperlink.Inlines.OfType<Run>().Single().Text, Is.EqualTo(url));
+            Assert.That(hyperlink.NavigateUri.ToString(), Is.EqualTo(url));
+        }
+    }
+}