+++ src/Plainion.Windows/Controls/Text/AutoCorrection/IAutoCorrection.cs
@@ -1,0 +1,10 @@
+﻿using System.Windows.Documents;
+
+namespace Plainion.Windows.Controls.Text.AutoCorrection
+{
+    public interface IAutoCorrection
+    {
+        bool TryApply(TextRange range);
+        TextPointer TryUndo(TextPointer start);
+    }
+}